refactored code to use routers.  also fixed creator->owner bug in schema
